Various fixes and updates:
- Exposed visibility phase rotation functions to Python interface.

- Allowed visibility time selection to work with irregularly sampled data in
Measurement Sets, which was not possible before. For this to work, removed
option to make time-snapshot images, as this concept isn't compatible
with the Measurement Set data format (and was less useful anyway).
Time-snapshot images can still be made by running the imager repeatedly with
a restricted time range if required.

- Simplified interface to oskar_imager_update().

- Improved safety of gridding functions to avoid potential integer overflows
for very (very) large images, and ensured all parts of single precision
gridding functions are done in single precision. Minor optimisation to grid
index calculations.

- Added imager run log to FITS headers.

- Added time taken for data reads and grid updates to run log.

- Moved FFT functions from imager to math folder.