First cut to write MSs with ADIOS2
These changes provide no conditional compilation, no cmake support, etc,
and they are slightly difficult to compile in systems where dependencies
are not found in standard locations. Only the DATA column is actually
being written with ADIOS2, which is *probably* what we want, but I'm not
entirely sure of that either.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>