Reacting to early cuda runtime errors
In oskar_device_count the status flag was not being set when if a
problem was found during the call to cudaGetDeviceCount. That meant that
a call to oskar_device_count from oskar_interferometer_set_gpus failed
to recognize this problem, and behaved as if the call to
cudaGetDeviceCount had been successful (which wasn't, we were getting a
cudaErrorInsufficientDriver error when using cuda 10.0 in one of our
testing platforms).

The oskar_cuda_system_info program handles this gracefully though,
because the underlying oskar_cuda_device_info_scan function does check
the return code from cudaGetDeviceCount.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>