Passing down MPI communicators to ADIOS/casacore
Both the Table (and related) and the Adios2StMan classes on casacore
allow passing down an explicit communicator to perform collective
operations. We are now allowing OSKAR to accept such communicators,
defaulting to MPI_COMM_WORLD if nothing is given.

I've also renamed the adios2-specific method to be better aligned with
the original one.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>