Avoid CUDA interactions by default
In more than one platform already we have hit incompatibility problems
between the NVidia kernel driver and the installed CUDA runtime. When
trying to workaround these errors by using the CPU version of OSKAR it
has been impossible because by default OSKAR will always query the
available CUDA devices, even when not required.

This patch solves that situation. It defaults the construction of the
OSKAR interferometer to *not* use GPUs, and when setting GPUs it only
queries available devices when the required number of devices is
different from 0.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>