MPI communicator and ADIOS2 usage are independent
In the previous iteration of this code we only passed down the
communicator when ADIOS2 was being used. This left us unable to use the
standard storage with an explicit MPI communicator, as it would still
default to using the MPI_COMM_WORLD internally.

With this change I'm changing this situation so the MPI communicator
passing is always performed when a valid communicator is passed down;
additionally, if ADIOS2 is requested then we setup the storage manager
for the DATA column. Both things are independent, with the exception
that ADIOS2 requires an MPI communicator. This new view of making the
MPI communicator the more important parameter, and the usage of ADIOS2
an option to toggle when MPI is in use, is now made explicit in the
structure and naming of the functions exported by the OSKAR ms library.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>